feat: Implement automatic site version checking
Adds logic to periodically check for a new version of the site by fetching `version.txt`. If a new version is detected, it updates the `site-version` in local storage and reloads the page to ensure users are on the latest build. This helps prevent users from seeing outdated content due to caching.